# Conflicts:
#	prisma/migrations/20250120135313_add_available_shapes_sizes/migration.sql

+++ prisma/migrations/20250120135313_add_available_shapes_sizes/migration.sql
@@ -4,7 +4,7 @@ CREATE TABLE "AvailableShapesSizes" (
     "shape_name" TEXT NOT NULL,
     "height" NUMERIC,
     "width" NUMERIC ,
-    "additional_price" NUMERIC,
+    "additional_price" INTEGER,
     "image" TEXT,
 
     CONSTRAINT "AvailableShapesSizes_pkey" PRIMARY KEY ("id")
